+++ test/command_checker/indexes_test.rb
@@ -55,6 +55,14 @@ module CommandChecker
       assert_safe AddIndexConcurrently
     end
 
+    def test_add_index_concurrently_corruption
+      OnlineMigrations::Utils.stub(:developer_env?, false) do
+        with_target_version(14.3) do
+          assert_unsafe AddIndexConcurrently, "can cause silent data corruption"
+        end
+      end
+    end
+
     class AddIndexNewTable < TestMigration
       def change
         create_table :users_new do |t|